moved generic UI component and modified array
